requirements.txt
requests
tqdm
python-dotenv
pandas
duckdb